cfg: move tests to separate file
that way we don't have to re-check the entire project when a test is
changed
